+++ app/sections/image-gallery/index.tsx
@@ -1,15 +1,12 @@
 import type {HydrogenComponentSchema} from '@weaverse/hydrogen';
 import {forwardRef} from 'react';
 import type {SectionProps} from '~/sections/shared/Section';
-import {Section, sectionConfigs} from '~/sections/shared/Section';
+import {Section, sectionInspector} from '~/sections/shared/Section';
 
-type ImageGalleryProps = SectionProps & {
-  heading: string;
-  description: string;
-};
+type ImageGalleryProps = SectionProps;
 
 let ImageGallery = forwardRef<HTMLElement, ImageGalleryProps>((props, ref) => {
-  let {heading, description, children, ...rest} = props;
+  let {children, ...rest} = props;
   return (
     <Section ref={ref} {...rest}>
       {children}
@@ -21,20 +18,20 @@ export default ImageGallery;
 
 export let schema: HydrogenComponentSchema = {
   type: 'image-gallery',
-  title: 'Image Gallery',
-  childTypes: ['heading', 'description', 'image-gallery--items'],
-  inspector: [sectionConfigs],
+  title: 'Image gallery',
+  childTypes: ['subheading', 'heading', 'description', 'image-gallery--items'],
+  inspector: [sectionInspector],
   toolbar: ['general-settings', ['duplicate', 'delete']],
   presets: {
     children: [
       {
         type: 'heading',
-        content: 'Hello World from Weaverse!',
+        content: 'Image Gallery',
       },
       {
         type: 'description',
         content:
-          "Stay on trend this season with our newest arrivals. We've curated the latest looks in women's fashion so you can refresh your wardrobe with ease. From flowy dresses and printed blouses perfect for summer, to cozy sweaters and tall boots for cooler months, our new collection has versatile pieces to take you from work to weekend. Shop cute and comfortable athleisure wear, chic handbags, and so much more.",
+          'Showcase your chosen images. This visual focus will enhance user engagement and understanding of your offerings.',
       },
       {
         type: 'image-gallery--items',